# docs/swagger.yaml
basePath: /
definitions:
  model.Blog:
    properties:
      content:
        description: 内容
        type: string
      createdAt:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      flag:
        description: 标记
        type: string
      id:
        description: 主键
        type: integer
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      title:
        description: 标题
        type: string
      typeId:
        description: 分类id
        type: integer
      typeName:
        description: 分类类名
        type: string
      updatedAt:
        description: 更新时间
        type: integer
      userId:
        description: 用户id
        type: integer
      username:
        description: 用户名
        type: string
      views:
        description: 浏览次数
        type: integer
    type: object
  model.Tag:
    properties:
      blogs:
        items:
          $ref: '#/definitions/model.Blog'
        type: array
      id:
        description: 主键
        type: integer
      tagName:
        description: 标签名
        type: string
    type: object
  model.Type:
    properties:
      amount:
        type: integer
      blogs:
        items:
          $ref: '#/definitions/model.Blog'
        type: array
      id:
        type: integer
      typeName:
        type: string
    type: object
  model.User:
    properties:
      avatar:
        description: 头像
        type: string
      blogs:
        items:
          $ref: '#/definitions/model.Blog'
        type: array
      createdAt:
        description: 创建时间
        type: integer
      email:
        description: 邮箱 (做登录账户)
        type: string
      id:
        description: 主键
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
      username:
        description: 用户名
        type: string
      uuid:
        description: 用户UUID
        type: string
    type: object
  request.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.Post:
    properties:
      content:
        type: string
      description:
        maxLength: 100
        minLength: 10
        type: string
      flag:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 15
        minLength: 1
        type: string
      typeId:
        type: integer
      userId:
        type: integer
      username:
        type: string
    required:
    - content
    - flag
    - typeId
    type: object
  request.Register:
    properties:
      checkPwd:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  response.PageList:
    properties:
      dataList: {}
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.User:
    properties:
      avatar:
        type: string
      id:
        type: integer
      username:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
  description: This is blog server.
  title: Blog API Swagger
  version: "1.0"
paths:
  /archive/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 按年份显示全部博客信息
      tags:
      - Example/Type
  /blog/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除的博客id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 个人博客删除
      tags:
      - Example/Blog
    get:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取博客信息
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取博客信息
      tags:
      - Example/Blog
  /blog/latestList:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
      summary: 最新推荐
      tags:
      - Example/Blog
  /blog/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 每页显示的
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNo
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageList'
      security:
      - X-Token: []
      summary: 个人博客展示
      tags:
      - Example/Blog
  /blog/pageList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 每页显示的
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNo
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageList'
      summary: 博客展示
      tags:
      - Example/Blog
  /blog/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Blog
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/request.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 新增博客
      tags:
      - Example/Blog
  /blog/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Blog
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/request.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 修改博客
      tags:
      - Example/Blog
  /blog/update/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取博客信息
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - X-Token: []
      summary: 获取编辑的博客信息
      tags:
      - Example/Blog
  /search/blog:
    get:
      consumes:
      - application/json
      parameters:
      - description: 标题
        in: query
        name: query
        required: true
        type: string
      - description: 每页显示
        in: query
        name: pageSize
        type: integer
      - description: 页数
        in: query
        name: pageNo
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageList'
      summary: 查询所有博客
      tags:
      - System/Search
  /search/priblog:
    get:
      consumes:
      - application/json
      parameters:
      - description: 标题
        in: query
        name: title
        type: string
      - description: 起始时间
        in: query
        name: startDate
        type: string
      - description: 结束时间
        in: query
        name: endDate
        type: string
      - description: 每页显示
        in: query
        name: pageSize
        type: integer
      - description: 页数
        in: query
        name: pageNo
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageList'
      security:
      - X-Token: []
      summary: 查询个人博客
      tags:
      - System/Search
  /tag/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: 标签列表
      tags:
      - Example/Tag
  /type/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Type'
            type: array
      summary: 分类展示(不排序)
      tags:
      - Example/Type
  /type/listOrder:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Type'
            type: array
      summary: 分类展示(amount降序)
      tags:
      - Example/Type
  /type/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分类id
        in: query
        name: id
        required: true
        type: integer
      - description: 每页显示的
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNo
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageList'
      summary: 分类查询分页
      tags:
      - Example/Type
  /type/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Type
        in: body
        name: blog
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 新增分类
      tags:
      - Example/Type
  /upload/mdFile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 图片上传
      tags:
      - System/Upload
  /user/info:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.User'
      security:
      - X-Token: []
      summary: 获取用户信息
      tags:
      - System/User
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 登录
      tags:
      - System/User
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 注册
      tags:
      - System/User
  /user/updateName:
    put:
      consumes:
      - application/json
      parameters:
      - description: update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - X-Token: []
      summary: 修改用户名
      tags:
      - System/User
securityDefinitions:
  X-Token:
    in: heaader
    name: X-Token
    type: apiKey
swagger: "2.0"